QR scanning fix
