+++ lib/views/after_auth_screens/profile/profile_page.dart
@@ -1,10 +1,17 @@
 import 'package:flutter/material.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:qr_flutter/qr_flutter.dart';
+import 'package:social_share/social_share.dart';
+import 'package:talawa/custom_painters/telegram_logo.dart';
+import 'package:talawa/custom_painters/whatsapp_logo.dart';
 import 'package:talawa/enums/enums.dart';
 import 'package:talawa/locator.dart';
 import 'package:talawa/models/options/options.dart';
+import 'package:talawa/services/graphql_config.dart';
 import 'package:talawa/services/size_config.dart';
 import 'package:talawa/utils/app_localization.dart';
 import 'package:talawa/view_model/after_auth_view_models/profile_view_models/profile_page_view_model.dart';
+import 'package:talawa/view_model/lang_view_model.dart';
 import 'package:talawa/views/base_view.dart';
 import 'package:talawa/widgets/custom_avatar.dart';
 import 'package:talawa/widgets/custom_list_tile.dart';
@@ -132,8 +139,8 @@ class _ProfilePageState extends State<ProfilePage> {
                                         .strictTranslate(
                                             'Help us to develop for you'),
                                   ),
-                                  onTapOption: () => model.modalBottomSheet(
-                                      context, ModalSheet.donation)),
+                                  onTapOption: () =>
+                                      modalBottomSheet(context, donate())),
                               CustomListTile(
                                   key: const Key('Option3'),
                                   index: 3,
@@ -149,8 +156,8 @@ class _ProfilePageState extends State<ProfilePage> {
                                     subtitle: AppLocalizations.of(context)!
                                         .strictTranslate('Invite to org'),
                                   ),
-                                  onTapOption: () => model.modalBottomSheet(
-                                      context, ModalSheet.invite)),
+                                  onTapOption: () => modalBottomSheet(
+                                      context, invite(context))),
                               CustomListTile(
                                   key: const Key('Option4'),
                                   index: 3,
@@ -176,5 +183,114 @@ class _ProfilePageState extends State<ProfilePage> {
                   ),
           );
         });
+  }
+
+  Widget invite(BuildContext context) {
+    final String url =
+        'https://cyberwake.github.io/applink/invite?selectLang=${AppLanguage().appLocal.languageCode}&setUrl=${GraphqlConfig.orgURI}&selectOrg=${userConfig.currentOrg.id!}';
+    final String qrData =
+        '${GraphqlConfig.orgURI}?orgid=${userConfig.currentOrg.id!}';
+    print(url);
+    print(qrData);
+    return Column(
+      mainAxisSize: MainAxisSize.max,
+      mainAxisAlignment: MainAxisAlignment.center,
+      crossAxisAlignment: CrossAxisAlignment.center,
+      children: [
+        QrImage(
+          data: qrData,
+          version: QrVersions.auto,
+          size: 200.0,
+          foregroundColor: Colors.black,
+        ),
+        SizedBox(
+          height: SizeConfig.screenHeight! * 0.08,
+        ),
+        Text(
+          'Scan the QR to join ${userConfig.currentOrg.name}',
+          style: const TextStyle(color: Colors.black),
+        ),
+        SizedBox(
+          height: SizeConfig.screenHeight! * 0.02,
+        ),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            iconButton(
+                const FaIcon(
+                  FontAwesomeIcons.twitter,
+                  size: 35,
+                  color: Color(0xFF1DA1F2),
+                ),
+                () async => SocialShare.shareTwitter('Join us', url: url)),
+            iconButton(
+                CustomPaint(
+                  size: Size(
+                      50,
+                      (50 * 1.004)
+                          .toDouble()), //You can Replace [WIDTH] with your desired width for Custom Paint and height will be calculated automatically
+                  painter: WhatsappLogo(),
+                ),
+                () async => SocialShare.shareWhatsapp(url)),
+            iconButton(
+                CustomPaint(
+                  size: Size(
+                      45,
+                      (45 * 1)
+                          .toDouble()), //You can Replace [WIDTH] with your desired width for Custom Paint and height will be calculated automatically
+                  painter: TelegramLogo(),
+                ),
+                () async => SocialShare.shareTelegram(url)),
+            iconButton(
+                const FaIcon(
+                  FontAwesomeIcons.shareAlt,
+                  size: 30,
+                  color: Color(0xff40c351),
+                ),
+                () async => SocialShare.shareOptions(url)),
+          ],
+        )
+      ],
+    );
+  }
+
+  Widget iconButton(Widget icon, Function onTap) {
+    return Stack(
+      children: [
+        IconButton(
+            onPressed: () {
+              print('tapped');
+              onTap();
+            },
+            icon: icon),
+      ],
+    );
+  }
+
+  void modalBottomSheet(BuildContext context, Widget child) {
+    showModalBottomSheet(
+        context: context,
+        shape: const RoundedRectangleBorder(
+          borderRadius: BorderRadius.only(
+              topLeft: Radius.circular(30), topRight: Radius.circular(30)),
+        ),
+        builder: (BuildContext context) {
+          return ClipRRect(
+            borderRadius: const BorderRadius.only(
+                topLeft: Radius.circular(30), topRight: Radius.circular(30)),
+            child: Container(
+              height: MediaQuery.of(context).size.height * 0.75,
+              decoration: const BoxDecoration(
+                color: Colors.white,
+              ),
+              child: child,
+            ),
+          );
+        });
+  }
+
+  Widget donate() {
+    return Container();
   }
 }
