+++ pubspec.yaml
@@ -38,6 +38,7 @@ dependencies:
   social_share: ^2.1.1
   timelines: ^0.1.0
   uni_links: ^0.5.1
+  vibration: ^1.7.4-nullsafety.0
   visibility_detector: ^0.2.0
 
 dev_dependencies:
